+++ src/common/components/Survey/SubjectiveAnswer/SubjectiveAnswser.server.tsx
@@ -1,4 +1,4 @@
-import ChartContainer from '@/common/components/Survey/ChartContainer/ChartContainer';
+import ChartContainer from '@common/components/Survey/ChartContainer/ChartContainer.server';
 
 import { SubjectiveAnswersProps } from './SubjectiveAnswer.types';
